Move configuration to its own class
Creates a new configuration class that should be backwards compatiable with
existing setups. When the module is loaded, it will initialize a Configuration object
and stores it in the class variable @@configuration. This configuration object
attempts to load variables based on the app environment in a way that is similar
to what we had previously.

Externally facing module methods remain similar, the most pertinent being
```load_config(path = nil, environment = nil)``` and ```config(key)```.

Adds a getter for the class variable via configuration.